extracted the send email function to the external function
